requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
pillow
nltk
rouge-score
transformers
yacs
pyyaml
tqdm
git+https://github.com/openai/CLIP.git
hi-ml-multimodal
